Create a navigation page for insights, so each tool can live on a single-purpose page.
